requirements.txt
Flask
Flask-Login
Flask-PyMongo
Flask-Bcrypt
python-dotenv
google-generativeai 